requirements.txt
pytesseract
Pillow
discord.py
